backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.2.0
pydantic==1.10.13
langgraph==0.0.20
langchain==0.1.0
langchain-openai==0.0.2
openai>=1.6.1
finnhub-python==2.4.18
python-dotenv==1.0.0
httpx==0.25.2
pandas==2.1.4
numpy>=1.26.0
